Fan Speed fix for Air Purifier Pro
This is a bug fix for Air Purifier Pro

Adjusting the FAN SPEED in MANUAL mode in the Home App, the FAN SPEED slider has values from 0-100. This plugin translates the slider setting to a number between 0 and 10, so sending out this parsed value to the air purifier. I assume the reason is that other Xiaomi Air Purifiers using fan speed settings from 0 - 10. However, Air Purifier Pro supports 17 different values for the FAN SPEED (in Mi App the minimum setting is 7~12m2 and the maximum is 35~60m2). All together there are 17 different values available. Because the plugin limit is 10, so the maximum FAN SPEED setting stops around ~60%. When the plugin checks for the actual fan speed to adjust the slider position, the same incorrect calculation results in setting the slider to the lower range instead of the actual speed.

I used a different formula to translate the slider setting to a value between 1 and 17.